Variable renaming using UDDU web + support for for more nodes
Refactoring for ReachingDefinitionsAnalyis and VariableNameCollector.
Decided to keep all the cases within the tame IR ecosystem and not use
caseNameExpr, therefore I have separate cases for for, while and if
statements. This way eases handling the nodes(I ran into hierachy and
casting issues using NameExpr). Support for for statement is there,
seems sound but I was not able to get any of the traversers to reach
the case.
Ran several benchmarks that highlighted new blind spots in my analysis
mainly statements involving arrays. Will look into it asap.