Fix pretty printer bug.
The first argument to String.format is a format string. If we pass in
a literal string with no arguments, and that string happens to contain
a %s (which e.g. a string literal expression could), it'll give an
error. So have wrap() just use the string directly if no arguments
are given.