Add basic inline variable refactoring.
This required almost no change to the layout preserving code. The main
thing is that if we inline a binary expression into another binary
expression, then the precedence rules might be messed up, so to be
safe when a node is copied (another use for transformer.copy), we wrap
binary expressions in a synthesized pair of parens.