Refactor MScriptInliner to extend Refactoring.
This is a big commit because this was the first refactoring I tried that
needed multiple files (the script to inline, and all the other files
that might call that script). This forced a few things (notably
RefactoringContext which is used in a lot of places) to change so that
they weren't tied to one MatlabProgram -- just a Project. The
RefactoringContext also knows whether transformations should be
layout-preserving, and exposes a method to get the correct Transformer
for a given AST node or MatlabProgram, instead of relying on the right
Transformer being passed in (because the Transformer is tied to a
specific MatlabProgram).