Modified main
Added a condition to ignore pForFunc. The pForFunc is a special function used to specify the number of shared variables in a parallel for. It is always called before the parfor loop. Since tamer expects a function definition to go along with it a dummy function has to be created. However it should not be parsed into vrir.